refactor: create LoginForm component
I moved most code in pages/login.vue to a separate LoginForm component.
The intention is to have a separate vue storybook for the LoginForm
component.
